use more size_t, int32_t

+++ sperl_constant_pool.h
@@ -6,13 +6,13 @@
 // Array
 struct SPerl_constant_pool {
   SPerl_VALUE_T* values;
-  int64_t length;
-  int64_t capacity;
+  size_t length;
+  size_t capacity;
 };
 
 // Array function
 SPerl_CONSTANT_POOL* SPerl_CONSTANT_POOL_new(SPerl* sperl);
-void SPerl_CONSTANT_POOL_extend(SPerl* sperl, SPerl_CONSTANT_POOL* constant_pool, int64_t extend);
+void SPerl_CONSTANT_POOL_extend(SPerl* sperl, SPerl_CONSTANT_POOL* constant_pool, size_t extend);
 
 void SPerl_CONSTANT_POOL_push_int(SPerl* sperl, SPerl_CONSTANT_POOL* constant_pool, int32_t value);
 void SPerl_CONSTANT_POOL_push_long(SPerl* sperl, SPerl_CONSTANT_POOL* constant_pool, int64_t value);
